fixed caching issues
Added an intermediate cache in the upload tab - "toBeUploaded", that will be referenced for the thumbnail gallery in the upload page.

Following that, when the user clicks "Continue to Edit", the cached images gets appended into the "uploadedImages" cache. This will be the master cache that is referenced in all the other tabs.